requirements.txt
pydantic-ai>=0.6.0
pydantic>=2.4.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.22.0
pytest>=7.3.1
httpx>=0.24.1